+++ lib/src/uicolor.dart
@@ -1,0 +1,14 @@
+import 'package:malison/malison.dart';
+
+class UIColor {
+  static const playerAttack = Color.lightGray;
+  static const enemyAttack = Color.lightRed;
+  static const playerDie = Color.darkRed;
+  static const enemyDie = Color.darkOrange;
+
+  static const welcomeText = Color.lightAqua;
+
+  static const barText = Color.white;
+  static const barFilled = Color.green;
+  static const barEmpty = Color.darkRed;
+}